+++ index.js
@@ -3,8 +3,9 @@ Promise.config({
     cancellation: true, //used for canceling promise chain in case of early request response
 });
 
+exports = module.exports = require('./lib/service.js');
+
 exports.database = {
-    sequelizeBuilder: require('./lib/database/sequelize.js'),
     CouchbaseCluster: require('./lib/database/couchbase.js')
 };
 
@@ -22,13 +23,6 @@ exports.Router               = require('./lib/express/router.js');
 exports.Route                = require('./lib/express/route.js');
 exports.RequestType          = require('./lib/express/requestType.js');
 exports.moduleLoader         = require('./lib/moduleLoader.js');
-exports.objectFilter         = require('./lib/objectFilter.js');
-exports.serviceIntegrity     = require('./lib/serviceIntegrity.js');
-exports.staticData           = require('./lib/staticData');
-
-exports.models = {
-    odm: {}
-};
 
 exports.error       = {
     IncompleteDataError : require('./lib/error/incompleteDataError.js'),